implements version as double instead of int

+++ src/main/scala/io/mdcatapult/doclib/models/DoclibFlag.scala
@@ -18,7 +18,7 @@ object DoclibFlag  {
   */
 case class DoclibFlag(
                        key: String,
-                       version: Int,
+                       version: Double,
                        hash: String,
                        started: LocalDateTime,
                        ended: Option[LocalDateTime] = None,
